requirements.txt
python-telegram-bot==22.5
yt-dlp
requests
python-dotenv